+++ src/merge/output_control.h
@@ -171,7 +171,7 @@ extern std::string g_default_language;
 extern float g_video_fps;
 extern generic_packetizer_c *g_video_packetizer;
 
-extern bool g_write_cues, g_cue_writing_requested;
+extern bool g_write_cues, g_cue_writing_requested, g_write_date;
 extern bool g_no_lacing, g_no_linking, g_use_durations, g_no_track_statistics_tags;
 
 extern bool g_identifying;